+++ testsuites/psxtests/psxstat/test.c
@@ -197,7 +197,7 @@ void stat_a_file(
 
     printf("\n...st_dev     (0x%x:0x%x)\n", major1, minor1 );
     printf(  "...st_ino     %x\n", (int) statbuf.st_ino );
-    printf(  "...st_mode    %o\n", statbuf.st_mode );
+    printf(  "...st_mode    %o\n", (unsigned int) statbuf.st_mode );
     printf(  "...st_nlink   %x\n", statbuf.st_nlink );
     printf(  "...st_uid     %d\n", statbuf.st_uid );
     printf(  "...st_gid     %d\n", statbuf.st_gid );
